README
BIOVEG is an R script that covers common statistical analyses (basic and advanced) applied in plant ecology. Thus, the script is a compilation of methods based on several published R packages, being applied to the modeling and statistical tests of Brazilian Atlantic Forest and temperate German physiognomies. The main aim of this script is to share knowledge and support / facilitate the work of other researchers and graduate students. In order to achieve this goal, the script analyses are based on data sets provided by the authors. Therefore, the user can perform all analyses contained in the script in the R software using real data. Each of the analyses has descriptions about their applications and some of them also about their limitations and care. At the end of the script, literature on statistical topics is also suggested. In this third version of the script, we expanded the number of statistical methods and focused mainly on regression modeling. The authors intend to constantly update and implement the script with more analyses. If you have any problems downloading or using it, please contact us by email: eciodiniz@gmail.com